Renew StatisticImage and SSubmslee.cpp

+++ shared/SStatisticsImage.cpp
@@ -14,7 +14,7 @@
 //
 //	Revision number:		3.0
 //
-//	Revision date:			12/16/2016
+//	Revision date:			07/01/2016
 //
 //	Language:				C
 //
@@ -67,12 +67,10 @@
   
 #if defined multispec_win   
 	#include	"SMulSpec.h"    
-	#include	"WStatImageDlg.h" 
+//	#include	"WMosaicTwoImagesDialog.h" 
 	#include "SExtGlob.h"
-	#include "resource.h"
-
-	#define  kStatisticsImageStrID          172
-
+	#include "WStatImageDlg.h"
+#define  kStatisticsImageStrID          172
 #endif	// defined multispec_win
 
 #ifdef multispec_lin 
@@ -1770,7 +1768,7 @@ Boolean LoadStatisticsImageSpecs (
 // Called By:			StatisticsImageControl 
 //
 //	Coded By:			Larry L. Biehl			Date: 11/15/1991
-//	Revised By:			Larry L. Biehl			Date: 12/16/2016
+//	Revised By:			Larry L. Biehl			Date: 06/29/2016
 	
 Boolean StatisticsImageDialog (
 				FileInfoPtr							fileInfoPtr)
@@ -2131,7 +2129,7 @@ Boolean StatisticsImageDialog (
 			// Set default text selection to first edit text item						
 	
 	if (selectItem != 0)	
-		SelectDialogItemText (dialogPtr, selectItem, 0, INT16_MAX);
+		SelectDialogItemText (dialogPtr, selectItem, 0, INT_MAX);
 		
 	modalDone = FALSE;
 	itemHit = 0;
@@ -2269,7 +2267,7 @@ Boolean StatisticsImageDialog (
 					if (itemHit == 16)
 						{
 						ShowDialogItems (dialogPtr, 17, 20);
-						SelectDialogItemText (dialogPtr, 18, 0, INT16_MAX);
+						SelectDialogItemText (dialogPtr, 18, 0, INT_MAX);
 						
 						}		// end "if (itemHit == 16)"
 						
@@ -2305,7 +2303,7 @@ Boolean StatisticsImageDialog (
 					SetDLogControl (dialogPtr, 6, 0);
 					HideDialogItems (dialogPtr, 7, 12);
 					ShowDialogItems (dialogPtr, 21, 30);
-					SelectDialogItemText (dialogPtr, 25, 0, INT16_MAX);
+					SelectDialogItemText (dialogPtr, 25, 0, INT_MAX);
 					
 					break;
 						
@@ -2499,7 +2497,7 @@ Boolean StatisticsImageDialog (
 		END_CATCH_ALL
 	
 	#endif	// defined multispec_win  
-
+	
    
 #if defined multispec_lin
    try{
@@ -2615,14 +2613,14 @@ void StatisticsImageDialogInitialize (
 											25,
 											entireIconItem,
 											kDontAdjustToBaseImage);
-
+	
 	*localNumberClassesPtr = statisticsImageSpecsPtr->numberClasses;
 	
 			//	Channels to use. Make all channels the default													
 	
 	*channelSelectionPtr = statisticsImageSpecsPtr->channelSet;
 	*localNumberChannelsPtr = statisticsImageSpecsPtr->numberFeatures;
-  
+   
    //if (gProjectInfoPtr != NULL &&
 	//			gProjectInfoPtr->numberStatTrainClasses > 0 &&
 	//				gProjectInfoPtr->statisticsCode == kMeanCovariance)
@@ -2694,7 +2692,7 @@ void StatisticsImageDialogInitialize (
 		*maxNumberChannelsPtr = imageWindowInfoPtr->totalNumberChannels;
 									
 		}		// end "else gProjectInfoPtr == NULL || ..."
- 	
+	
 			// Set control bullet for "Use overall min/max".							
 			
 	*minMaxSettingCodePtr = statisticsImageSpecsPtr->minMaxSettingCode;
@@ -2710,7 +2708,7 @@ void StatisticsImageDialogInitialize (
 			
 	SetDLogControl (dialogPtr, IDC_UserSettingRadio,
 								(statisticsImageSpecsPtr->minMaxSettingCode == 3));
-							
+								
 	if (statisticsImageSpecsPtr->minMaxSettingCode == 3)
 		{
 		ShowDialogItem (dialogPtr, IDC_MinPrompt);
